remove UserAssignedToProjectDataSource  interface
