+++ src/main/kotlin/data/source/ProjectDataSource.kt
@@ -1,6 +1,7 @@
 package org.example.data.source
 
 import data.dto.ProjectDto
+import data.dto.UserAssignedToProjectDto
 
 interface ProjectDataSource {
     suspend fun addProject(project: ProjectDto): Boolean
@@ -8,4 +9,6 @@ interface ProjectDataSource {
     suspend fun deleteProject(projectToDelete: ProjectDto): Boolean
     suspend fun getAllProjects(): List<ProjectDto>
     suspend fun getProjectsByUsername(username: String): List<ProjectDto>
+
+    suspend fun getUsersAssignedToProjectByUserName(userName: String): List<UserAssignedToProjectDto>
 }
