+++ src/main/kotlin/data/source/local/UserCSVDataSource.kt
@@ -1,23 +1,22 @@
 package org.example.data.source.local
 
 import data.dto.MateTaskAssignmentDto
+import data.dto.UserAssignedToProjectDto
 import data.dto.UserDto
-import org.example.data.source.UserAssignedToProjectDataSource
 import org.example.data.source.UserDataSource
 import org.example.data.source.local.csv_reader_writer.IReaderWriter
 
 class UserCSVDataSource(
     private val userReaderWriter: IReaderWriter<UserDto>,
     private val mateTaskAssignmentReaderWriter: IReaderWriter<MateTaskAssignmentDto>,
-    private val userAssignedToProjectDataSource: UserAssignedToProjectDataSource,
-
+    private val userAssignedToProjectReaderWriter: IReaderWriter<UserAssignedToProjectDto>,
 ) : UserDataSource {
     override suspend fun addUser(user: UserDto): Boolean = userReaderWriter.append(listOf(user))
 
     override suspend fun getAllUsers(): List<UserDto> = userReaderWriter.read()
 
     override suspend fun getUsersByProjectId(projectId: String): List<UserDto> =
-        userAssignedToProjectDataSource.getUsersAssignedToProjectByProjectId(projectId)
+        getUsersAssignedToProjectByProjectId(projectId)
             .map { userAssignedToProject -> userAssignedToProject.username }
             .let { userIds -> getAllUsers().filter { user -> user.id in userIds } }
 
@@ -35,16 +34,28 @@ class UserCSVDataSource(
         getAllUsers().filterNot { mappedUser -> mappedUser.id == user.id }
             .let { updatedUsers -> userReaderWriter.overWrite(updatedUsers) }
 
-    override suspend fun deleteUserFromProject(projectId: String, username: String): Boolean =
-        userAssignedToProjectDataSource.deleteUserFromProject(userName = username, projectId = projectId)
-
-    override suspend fun addUserToProject(projectId: String, username: String): Boolean =
-        userAssignedToProjectDataSource.addUserToProject(userName = username, projectId = projectId)
-
     override suspend fun addUserToTask(username: String, taskId: String): Boolean =
         mateTaskAssignmentReaderWriter.append(listOf(MateTaskAssignmentDto(username, taskId)))
 
     override suspend fun deleteUserFromTask(username: String, taskId: String): Boolean =
         mateTaskAssignmentReaderWriter.read().filterNot { mateTaskAssignment -> mateTaskAssignment.taskId == taskId }
             .let { filteredMateTaskAssignment -> mateTaskAssignmentReaderWriter.overWrite(filteredMateTaskAssignment) }
+
+    override suspend fun addUserToProject(projectId: String, userName: String): Boolean =
+        userAssignedToProjectReaderWriter.append(
+            listOf(UserAssignedToProjectDto(projectId = projectId, username = userName))
+        )
+
+    override suspend fun deleteUserFromProject(projectId: String, userName: String): Boolean =
+        userAssignedToProjectReaderWriter.read().filterNot { userAssignedToProject ->
+            userAssignedToProject.projectId == projectId && userAssignedToProject.username == userName
+        }.let { updatedUserAssignedToProjectList ->
+            userAssignedToProjectReaderWriter.overWrite(updatedUserAssignedToProjectList)
+        }
+
+    override suspend fun getUsersAssignedToProjectByProjectId(projectId: String): List<UserAssignedToProjectDto> =
+        userAssignedToProjectReaderWriter.read().filter { userAssignedToProject ->
+            userAssignedToProject.projectId == projectId
+        }
+
 }
